Add files via upload
This notebook performs sentiment analysis on movie reviews from the IMDb dataset using Recurrent Neural Networks (RNNs). It covers data loading and preparation, model building with an LSTM layer, and training the model to classify reviews as positive or negative. The approach includes decoding reviews for interpretation, padding sequences for consistency, and evaluating the model’s performance on test data.